feat(markers): host-driven PlotMarkers — MarkerService, rendering, generators.v1 + preview grid/pan (#296)
## Summary

Integration PR for the **plot-markers** feature on the host side. Adds the full host-side marker infrastructure that the Anomaly Detector plugin (and future generator toolboxes) build on:

- **`pj.generators.v1` / `GeneratorsHostService`** backed by the **ObjectStore-only** marker engine (`MarkerService`): a generator submitted to the host is the host's to MAINTAIN — it re-runs whole-series on each matching commit and keeps its marker object topic fresh until removed.
- **Marker storage as ObjectStore objects** (no separate MarkerStore): per-series topic keys, multi-dataset/retention slots, marker topics hidden from the catalog.
- **Rendering** in `pj_plotting`: pills, hollow point, off-screen culling; per-curve visibility toggle.
- **Dialog-host overlay**: markers + file pickers in toolbox panels; the embedded `ChartPreviewWidget` preview.
- Readable Lua marker API (`createMarker` + `create{Vertical,Horizontal,Point,Band}Marker`); SDK submodule advanced for `PlotMarkers`.

## Latest in this push

- **`ChartPreviewWidget` grid + pan** (`pj_dialog_host`): the embedded toolbox preview chart gains a background grid (major dashed / minor dotted) + middle-drag and Ctrl+left-drag panning, alongside the existing rubber-band + wheel zoom. All gated behind `setZoomEnabled()` (now idempotent) so opt-out charts are pixel-identical. Domain-neutral host change — no plugin edit needed (toolboxes already call `setChartZoomEnabled()`). Mirrors `PlotWidgetBase`'s scheme; the lightweight Qwt widget stays independent of `pj_plotting`.

## Test plan

- `dialog_engine_test`, `panel_engine_test`, `widget_binding_test` — 3/3 green.
- `./build.sh` builds clean.
- Manual: toolbox preview shows a grid; wheel + left-drag zoom; middle-drag / Ctrl+left-drag pan; right-click steps zoom out — matching the Transform Editor's PlotWidget.

+++ pj_runtime/src/DataSourceRuntimeHost.cpp
@@ -756,7 +756,8 @@ bool DataSourceRuntimeHost::cbPushMessage(
     if (policy == sdk::ObjectIngestPolicy::kLazyObjectsEagerScalars) {
       sdk::PayloadView seed;
       if (payload_anchor != nullptr) {
-        seed = sdk::PayloadView{Span<const uint8_t>{payload.data, payload.size}, std::move(payload_anchor)};
+        seed = sdk::PayloadView{
+            Span<const uint8_t>{payload.data, static_cast<size_t>(payload.size)}, std::move(payload_anchor)};
       } else if (payload.size > 0) {
         // Anchorless transient buffer: the bytes die with this call, so the
         // seed needs its own copy.
